requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.7.0
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
cloud-init==23.4.4
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
faiss-cpu==1.8.0
Flask==3.0.2
flask-ngrok==0.0.25
frozenlist==1.4.1
fsspec==2024.3.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
implicit==0.7.2
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lightfm==1.17
lightning-utilities==0.11.0
MarkupSafe==2.1.5
more-itertools==8.10.0
multidict==6.0.5
netifaces==0.11.0
networkx==3.2.1
nmslib==2.1.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.0
packaging==24.0
pandas==1.5.3
pexpect==4.8.0
psutil==5.9.8
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pybind11==2.6.1
PyGObject==3.42.1
PyHamcrest==2.0.2
pyinotify==0.9.6
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
pytorch-lightning==1.9.5
pytz==2022.1
PyYAML==5.4.1
rectools==0.4.2
requests==2.25.1
scikit-learn==1.4.1.post1
scipy==1.12.0
SecretStorage==3.3.1
service-identity==18.1.0
simplejson==3.17.6
six==1.16.0
sos==4.5.6
ssh-import-id==5.11
systemd-python==234
threadpoolctl==3.3.0
torch==1.13.1
torchmetrics==1.3.2
tqdm==4.66.2
Twisted==22.1.0
typeguard==2.13.3
typing_extensions==4.10.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
Werkzeug==3.0.1
xkit==0.0.0
yarl==1.9.4
zipp==1.0.0
zope.interface==5.4.0
